+++ plugin/light/index.js
@@ -1,0 +1,6 @@
+module.exports = {
+  DirectionalLight: require('./directional'),
+  HemisphericLight: require('./hemispheric'),
+  PointLight: require('./point'),
+  SpotLight: require('./spot'),
+};