+++ plugin/mesh/abstract.js
@@ -1,0 +1,10 @@
+module.exports = {
+  mixins: [require('../entity')],
+
+  props: {
+    options: {
+      type: Object,
+      default: () => ({}),
+    }
+  },
+};