+++ plugin/util/index.js
@@ -1,10 +1,14 @@
-let { Vector3 } = require('babylonjs');
+let { Vector3, Color3 } = require('babylonjs');
 let { crypto } = window;
 
 let isFloat = value => Number.isFinite(n) && !Number.isNaN(n);
+let isBetween0and1 = value => value <= 1 && value >= 0;
+let isHexColor = value => /^#?([\da-fA-F]{3}){1,2}$/.test(value);
 
 module.exports = {
   isFloat,
+  isBetween0and1,
+  isHexColor,
 
   id(size = 12) {
     let buf = new Uint8Array(size);
@@ -21,6 +25,18 @@ module.exports = {
     default: () => Vector3.Zero(),
   },
 
+  color3: {
+    validator: value => value != null
+      && ((Array.isArray(value) && Array.every(isFloat) && Array.every(isBetween0and1))
+        || value instanceof Color3
+        || (typeof value === 'string' && isHexColor(value))
+        || (isFloat(value.r) && isBetween0and1(value.r)
+          && isFloat(value.g) && isBetween0and1(value.g)
+          && isFloat(value.b) && isBetween0and1(value.b))
+    ),
+    default: () => Color3.White(),
+  },
+
   toVec3(value) {
     if (value instanceof Vector3) {
       return value;
@@ -30,5 +46,19 @@ module.exports = {
     }
     let { x, y, z } = value;
     return new Vector3(x, y, z);
+  },
+
+  toColor3(value) {
+    if (value instanceof Color3) {
+      return value;
+    }
+    if (Array.isArray(value)) {
+      return Color3.FromArray(value);
+    }
+    if (typeof value === 'string') {
+      return Color3.FromHexString(value);
+    }
+    let { r, g, b } = value;
+    return new Color3(r, g, b);
   },
 };