+++ CHANGELOG.md
@@ -1,10 +1,20 @@
+<a name="0.3.0"></a>
+# [0.3.0](https://github.com/Beg-in/3d/compare/0.2.0...0.3.0) (2018-01-30)
+
+
+### Features
+
+* **light:** add lights ([bd6cf0b](https://github.com/Beg-in/3d/commit/bd6cf0b))
+
+
+
 <a name="0.2.0"></a>
 # [0.2.0](https://github.com/Beg-in/3d/compare/0.1.0...0.2.0) (2018-01-29)
 
 
 ### Features
 
-* **entity:** add entity and some basic meshes ([ebb4a76](https://github.com/Beg-in/3d/commit/ebb4a76))
+* **entity:** add entity and some basic meshes ([2cc7c33](https://github.com/Beg-in/3d/commit/2cc7c33))